client/requirements.txt
opencv-python-headless
python-socketio[client]
numpy
ultralytics
